Changed the ReferenceType to unsigned from unsigned long for 32 bit platforms; unsigned long is 64 bits on x86-64

+++ src/plat/linux32/lisptype.h
@@ -24,8 +24,8 @@
 // case a LispSubList references the same sequence of LispObjects!)
 #ifdef USE_LONG_REF_COUNTS
 #define ReferenceBits 32
-#define ReferenceType unsigned long
-#define ReferenceMax ((unsigned long)0xffffffff)
+#define ReferenceType unsigned
+#define ReferenceMax ((unsigned)0xffffffff)
 #else
 #define ReferenceBits 16
 #define ReferenceType unsigned short
